refactor(clients): deprecate RetrieveNames
Remove RetrieveNames and modify RetrieveAllBasicInfo to replace it. Remove sensitive information when retrieving basic information
