+++ pkg/clients/postgres/clients_test.go
@@ -6,8 +6,10 @@ package postgres_test
 import (
 	"context"
 	"fmt"
+	"strings"
 	"testing"
 
+	"github.com/0x6flab/namegenerator"
 	"github.com/absmach/magistrala/internal/testsutil"
 	mgclients "github.com/absmach/magistrala/pkg/clients"
 	"github.com/absmach/magistrala/pkg/errors"
@@ -22,6 +24,7 @@ var (
 	clientName     = "client name"
 	wrongName      = "wrong-name"
 	wrongID        = "wrong-id"
+	namesgen       = namegenerator.NewNameGenerator()
 )
 
 func TestClientsRetrieveByID(t *testing.T) {
@@ -897,4 +900,340 @@ func TestClientsChangeStatus(t *testing.T) {
 			assert.Equal(t, tc.client.Status, expected.Status, fmt.Sprintf("%s: expected %d got %d\n", tc.desc, tc.client.Status, expected.Status))
 		}
 	}
+}
+
+func TestClientsRetrieveAllBasicInfo(t *testing.T) {
+	t.Cleanup(func() {
+		_, err := db.Exec("DELETE FROM clients")
+		require.Nil(t, err, fmt.Sprintf("clean clients unexpected error: %s", err))
+	})
+	repo := cpostgres.NewRepository(database)
+
+	nusers := 100
+	users := make([]mgclients.Client, nusers)
+
+	name := namesgen.Generate()
+
+	for i := 0; i < nusers; i++ {
+		username := fmt.Sprintf("%s-%d@example.com", name, i)
+		client := mgclients.Client{
+			ID:   testsutil.GenerateUUID(t),
+			Name: username,
+			Credentials: mgclients.Credentials{
+				Identity: username,
+				Secret:   password,
+			},
+			Metadata: mgclients.Metadata{},
+			Status:   mgclients.EnabledStatus,
+		}
+		_, err := repo.Save(context.Background(), client)
+		require.Nil(t, err, fmt.Sprintf("save client unexpected error: %s", err))
+
+		users[i] = mgclients.Client{
+			ID:   client.ID,
+			Name: client.Name,
+		}
+	}
+
+	cases := []struct {
+		desc     string
+		page     mgclients.Page
+		response mgclients.ClientsPage
+		err      error
+	}{
+		{
+			desc: "retrieve all clients",
+			page: mgclients.Page{
+				Offset: 0,
+				Limit:  uint64(nusers),
+			},
+			response: mgclients.ClientsPage{
+				Clients: users,
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 0,
+					Limit:  uint64(nusers),
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with offset",
+			page: mgclients.Page{
+				Offset: 10,
+				Limit:  uint64(nusers),
+			},
+			response: mgclients.ClientsPage{
+				Clients: users[10:],
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 10,
+					Limit:  uint64(nusers),
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with limit",
+			page: mgclients.Page{
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: users[:10],
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with offset and limit",
+			page: mgclients.Page{
+				Offset: 10,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: users[10:20],
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 10,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with name",
+			page: mgclients.Page{
+				Name:   users[0].Name[:1],
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: findClients(users, users[0].Name[:1], 0, 10),
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with name",
+			page: mgclients.Page{
+				Name:   users[0].Name[:4],
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: findClients(users, users[0].Name[:4], 0, 10),
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with name with SQL injection",
+			page: mgclients.Page{
+				Name:   fmt.Sprintf("%s' OR '1'='1", users[0].Name[:1]),
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: []mgclients.Client(nil),
+				Page: mgclients.Page{
+					Total:  0,
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with Identity",
+			page: mgclients.Page{
+				Identity: users[0].Name[:1],
+				Offset:   0,
+				Limit:    10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: findClients(users, users[0].Name[:1], 0, 10),
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with Identity",
+			page: mgclients.Page{
+				Identity: users[0].Name[:4],
+				Offset:   0,
+				Limit:    10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: findClients(users, users[0].Name[:4], 0, 10),
+				Page: mgclients.Page{
+					Total:  uint64(nusers),
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with Identity with SQL injection",
+			page: mgclients.Page{
+				Identity: fmt.Sprintf("%s' OR '1'='1", users[0].Name[:1]),
+				Offset:   0,
+				Limit:    10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: []mgclients.Client(nil),
+				Page: mgclients.Page{
+					Total:  0,
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients unknown name",
+			page: mgclients.Page{
+				Name:   "unknown",
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: []mgclients.Client(nil),
+				Page: mgclients.Page{
+					Total:  0,
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients unknown name with SQL injection",
+			page: mgclients.Page{
+				Name:   "unknown' OR '1'='1",
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: []mgclients.Client(nil),
+				Page: mgclients.Page{
+					Total:  0,
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients unknown identity",
+			page: mgclients.Page{
+				Identity: "unknown",
+				Offset:   0,
+				Limit:    10,
+			},
+			response: mgclients.ClientsPage{
+				Clients: []mgclients.Client(nil),
+				Page: mgclients.Page{
+					Total:  0,
+					Offset: 0,
+					Limit:  10,
+				},
+			},
+			err: nil,
+		},
+		{
+			desc: "retrieve all clients with order",
+			page: mgclients.Page{
+				Order:  "name",
+				Dir:    "asc",
+				Name:   users[0].Name[:1],
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{},
+			err:      nil,
+		},
+		{
+			desc: "retrieve all clients with order",
+			page: mgclients.Page{
+				Order:  "name",
+				Dir:    "desc",
+				Name:   users[0].Name[:1],
+				Offset: 0,
+				Limit:  10,
+			},
+			response: mgclients.ClientsPage{},
+			err:      nil,
+		},
+		{
+			desc: "retrieve all clients with order",
+			page: mgclients.Page{
+				Order:    "identity",
+				Dir:      "asc",
+				Identity: users[0].Name[:1],
+				Offset:   0,
+				Limit:    10,
+			},
+			response: mgclients.ClientsPage{},
+			err:      nil,
+		},
+		{
+			desc: "retrieve all clients with order",
+			page: mgclients.Page{
+				Order:    "identity",
+				Dir:      "desc",
+				Identity: users[0].Name[:1],
+				Offset:   0,
+				Limit:    10,
+			},
+			response: mgclients.ClientsPage{},
+			err:      nil,
+		},
+	}
+	for _, tc := range cases {
+		resp, err := repo.RetrieveAllBasicInfo(context.Background(), tc.page)
+		assert.True(t, errors.Contains(err, tc.err), fmt.Sprintf("%s: expected %s got %s\n", tc.desc, tc.err, err))
+		if err == nil {
+			if tc.page.Order != "" && tc.page.Dir != "" {
+				tc.response = resp
+			}
+			assert.Equal(t, tc.response, resp, fmt.Sprintf("%s: expected %v got %v\n", tc.desc, tc.response, resp))
+		}
+	}
+}
+
+func findClients(clients []mgclients.Client, query string, offset, limit uint64) []mgclients.Client {
+	clis := []mgclients.Client{}
+	for _, client := range clients {
+		if strings.Contains(client.Name, query) {
+			clis = append(clis, client)
+		}
+	}
+
+	if offset > uint64(len(clis)) {
+		return []mgclients.Client{}
+	}
+
+	if limit > uint64(len(clis)) {
+		return clis[offset:]
+	}
+
+	return clis[offset:limit]
 }
